# api/openapi.json
{
    "openapi": "3.0.3",
    "info": {
        "title": "db-adapter",
        "version": "0.0.1"
    },
    "servers": [
        {
            "description": "production server",
            "url": "http://0.0.0.0:7689"
        },
        {
            "description": "staging server",
            "url": "http://0.0.0.0:7699"
        },
        {
            "description": "development server",
            "url": "http://0.0.0.0:7698"
        }
    ],
    "paths": {
...
        "/person": {
            "post": {
                "summary": "Create a new person",
                "operationId": "createPerson",
                "parameters": [
                    {
                        "name": "X-User-ID",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/PersonRegistration"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Person created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Person"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "msg": {
                                            "type": "string"
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/person/google/{google_id}": {
            "post": {
                "summary": "Create a new person by Google ID",
                "operationId": "createPersonByGoogleId",
                "parameters": [
                    {
                        "name": "google_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/PersonRegistration"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Person created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Person"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "msg": {
                                            "type": "string"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "msg": {
                                            "type": "string"
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "patch": {
                "summary": "Create a new person by Google ID with invite code",
                "operationId": "createPersonByGoogleIdAndInviteCode",
                "parameters": [
                    {
                        "name": "google_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "required": [
                                    "invite_code",
                                    "person"
                                ],
                                "properties": {
                                    "invite_code": {
                                        "type": "string"
                                    },
                                    "person": {
                                        "$ref": "#/components/schemas/PersonRegistration"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Person created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Person"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "msg": {
                                            "type": "string"
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "get": {
                "summary": "Get a person by Google ID",
                "operationId": "getPersonByGoogleId",
                "parameters": [
                    {
                        "name": "google_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Person retrieved",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Person"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "msg": {
                                            "type": "string"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "msg": {
                                            "type": "string"
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/family-tree": {
            "get": {
                "summary": "Get family tree by person ID",
                "operationId": "getFamilyTreeById",
                "parameters": [
                    {
                        "name": "X-User-ID",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Family tree retrieved",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/FamilyTree"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "msg": {
                                            "type": "string"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "msg": {
                                            "type": "string"
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/family-tree-with-spouses": {
            "get": {
                "summary": "Get family tree by person ID with spouses included",
                "operationId": "getFamilyTreeWithSpousesById",
                "parameters": [
                    {
                        "name": "X-User-ID",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Family tree retrieved",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/FamilyTree"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "msg": {
                                            "type": "string"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "msg": {
                                            "type": "string"
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "PersonProperties": {
                "type": "object",
                "properties": {
                    "invite_code": {
                        "type": "string",
                        "nullable": true
                    },
                    "google_id": {
                        "type": "string",
                        "nullable": true
                    },
                    "first_name": {
                        "type": "string",
                        "nullable": true
                    },
                    "middle_name": {
                        "type": "string",
                        "nullable": true
                    },
                    "last_name": {
                        "type": "string",
                        "nullable": true
                    },
                    "titles": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "nullable": true
                    },
                    "suffixes": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "nullable": true
                    },
                    "extra_names": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "nullable": true
                    },
                    "aliases": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "nullable": true
                    },
                    "mothers_first_name": {
                        "type": "string",
                        "nullable": true
                    },
                    "mothers_last_name": {
                        "type": "string",
                        "nullable": true
                    },
                    "biological_sex": {
                        "type": "string",
                        "enum": [
                            "male",
                            "female",
                            "intersex",
                            "unknown",
                            "other"
                        ],
                        "nullable": true
                    },
                    "born": {
                        "type": "string",
                        "format": "date",
                        "nullable": true
                    },
                    "place_of_birth": {
                        "type": "string",
                        "nullable": true
                    },
                    "died": {
                        "type": "string",
                        "format": "date",
                        "nullable": true
                    },
                    "place_of_death": {
                        "type": "string",
                        "nullable": true
                    },
                    "life_events": {
                        "type": "array",
                        "items": {
                            "type": "object",
                            "properties": {
                                "from": {
                                    "type": "string",
                                    "format": "date"
                                },
                                "to": {
                                    "type": "string",
                                    "format": "date"
                                },
                                "description": {
                                    "type": "string"
                                }
                            }
                        },
                        "nullable": true
                    },
                    "occupations": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "nullable": true
                    },
                    "occupation_to_display": {
                        "type": "string",
                        "nullable": true
                    },
                    "limit": {
                        "type": "integer",
                        "nullable": true
                    },
                    "photos": {
                        "type": "array",
                        "items": {
                            "type": "object",
                            "properties": {
                                "url": {
                                    "type": "string"
                                },
                                "description": {
                                    "type": "string"
                                },
                                "date": {
                                    "type": "string",
                                    "format": "date"
                                },
                                "name": {
                                    "type": "string"
                                }
                            }
                        },
                        "nullable": true
                    },
                    "videos": {
                        "type": "array",
                        "items": {
                            "type": "object",
                            "properties": {
                                "url": {
                                    "type": "string"
                                },
                                "description": {
                                    "type": "string"
                                },
                                "date": {
                                    "type": "string",
                                    "format": "date"
                                },
                                "name": {
                                    "type": "string"
                                }
                            }
                        },
                        "nullable": true
                    },
                    "audios": {
                        "type": "array",
                        "items": {
                            "type": "object",
                            "properties": {
                                "url": {
                                    "type": "string"
                                },
                                "description": {
                                    "type": "string"
                                },
                                "date": {
                                    "type": "string",
                                    "format": "date"
                                },
                                "name": {
                                    "type": "string"
                                }
                            }
                        },
                        "nullable": true
                    },
                    "profile_picture": {
                        "type": "string",
                        "nullable": true
                    },
                    "verified": {
                        "type": "boolean",
                        "nullable": true
                    },
                    "email": {
                        "type": "string",
                        "nullable": true
                    },
                    "phone": {
                        "type": "string",
                        "nullable": true
                    },
                    "residence": {
                        "type": "object",
                        "properties": {
                            "city": {
                                "type": "string"
                            },
                            "country": {
                                "type": "string"
                            },
                            "zip_code": {
                                "type": "string"
                            },
                            "address_line_1": {
                                "type": "string"
                            },
                            "address_line_2": {
                                "type": "string"
                            }
                        },
                        "nullable": true
                    },
                    "religion": {
                        "type": "string",
                        "nullable": true
                    },
                    "baptized": {
                        "type": "string",
                        "nullable": true
                    },
                    "ideologies": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "nullable": true
                    },
                    "blood_type": {
                        "type": "string",
                        "nullable": true
                    },
                    "allergies": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "nullable": true
                    },
                    "medications": {
                        "type": "array",
                        "nullable": true,
                        "items": {
                            "type": "object",
                            "properties": {
                                "name": {
                                    "type": "string"
                                },
                                "description": {
                                    "type": "string",
                                    "nullable": true
                                },
                                "components": {
                                    "type": "string",
                                    "nullable": true
                                },
                                "dosage": {
                                    "type": "string",
                                    "nullable": true
                                },
                                "from": {
                                    "type": "string",
                                    "format": "date",
                                    "nullable": true
                                },
                                "to": {
                                    "type": "string",
                                    "format": "date",
                                    "nullable": true
                                }
                            }
                        }
                    },
                    "medical_conditions": {
                        "type": "array",
                        "items": {
                            "type": "object"
                        },
                        "nullable": true
                    },
                    "height": {
                        "type": "number",
                        "nullable": true
                    },
                    "weight": {
                        "type": "number",
                        "nullable": true
                    },
                    "hair_colour": {
                        "type": "string",
                        "nullable": true
                    },
                    "skin_colour": {
                        "type": "string",
                        "nullable": true
                    },
                    "eye_colour": {
                        "type": "string",
                        "nullable": true
                    },
                    "sports": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "nullable": true
                    },
                    "hobbies": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "nullable": true
                    },
                    "interests": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "nullable": true
                    },
                    "languages": {
                        "type": "array",
                        "items": {
                            "type": "object",
                            "properties": {
                                "language": {
                                    "type": "string"
                                },
                                "level": {
                                    "type": "string",
                                    "nullable": true
                                }
                            }
                        },
                        "nullable": true
                    },
                    "notes": {
                        "type": "array",
                        "items": {
                            "type": "object",
                            "properties": {
                                "date": {
                                    "type": "string",
                                    "format": "date",
                                    "nullable": true
                                },
                                "title": {
                                    "type": "string",
                                    "nullable": true
                                },
                                "note": {
                                    "type": "string"
                                },
                                "url": {
                                    "type": "string",
                                    "nullable": true
                                }
                            }
                        },
                        "nullable": true
                    }
                }
            },
            "PersonRegistration": {
                "type": "object",
                "required": [
                    "first_name",
                    "last_name",
                    "born",
                    "limit",
                    "mothers_first_name",
                    "mothers_last_name"
                ],
                "properties": {
                    "first_name": {
                        "type": "string"
                    },
                    "last_name": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "nullable": true
                    },
                    "limit": {
                        "type": "integer"
                    },
                    "biological_sex": {
                        "type": "string",
                        "enum": [
                            "male",
                            "female",
                            "intersex",
                            "unknown",
                            "other"
                        ]
                    },
                    "born": {
                        "type": "string",
                        "format": "date"
                    },
                    "mothers_first_name": {
                        "type": "string"
                    },
                    "mothers_last_name": {
                        "type": "string"
                    }
                }
            },
            "Person": {
                "type": "object",
                "properties": {
                    "Id": {
                        "type": "integer"
                    },
                    "ElementId": {
                        "type": "string"
                    },
                    "Labels": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    },
                    "Props": {
                        "$ref": "#/components/schemas/PersonProperties"
                    }
                }
            },
...
            "FamilyTree": {
                "type": "object",
                "properties": {
                    "people": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/OptimizedPersonNode"
                        }
                    },
                    "relationships": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/dbtypeRelationship"
                        }
                    }
                }
            },
...
        }
    }
}